Add a search page
The page uses the search endpoint in the python backend to
search all entries in the provided company data which match
the search terms. Clicking on the resultant entries then shows
the user the analytics of that particular entry, based on the data
provided.

Note: The s_no attribute of a search entry is then used to make
the call to calculate analytics of that particular row.

TODO: remove localhost from request origin
Implement the analytics display page.